requirements.txt
requests==2.32.3
python-dotenv==1.1.0
Django==5.2.3
djangorestframework==3.16.0
pymongo==4.13.2

